feat(security,deploy): T056 clang-tidy security_scan + T057 deployment docs
T056-001: Add security_scan CMake custom target (run-clang-tidy-18) with
  cert-*, bugprone-*, cppcoreguidelines-*, clang-analyzer-security.* checks
  over all library sources in src/main/cpp/isched/
T056-002: Create .clang-tidy at repo root with security-focused check set;
  each suppression documented with rationale (crash-prone checks, C API
  mandated patterns, Boost.Asio idioms, httplib raw-pointer API)
T056-003: Verify security_scan exits clean (EXIT:0, zero user-code warnings);
  fix findings found during scan:
  - bugprone-use-after-move in Server.cpp (real bug: ctx moved then accessed)
  - cert-err33-c: std::ignore strftime/snprintf in GqlExecutor + AuthMiddleware
  - cert-err58-cpp: NOLINTNEXTLINE on thread_local nlohmann::json in GqlExecutor
  - clang-diagnostic-unused-variable: remove dead request_counter/error_counter
  - clang-diagnostic-unused-lambda-capture: remove unused 'this' from 3 lambdas
  - clang-diagnostic-pessimizing-move: remove std::move on temporary
  - clang-diagnostic-defaulted-function-deleted: = default -> = delete for move
  - bugprone-misplaced-widening-cast: fix enc_len+final_len overflow in CryptoUtils
  - clang-diagnostic-unused-parameter: [[maybe_unused]] in DatabaseManager/config
  - bugprone-unused-local-non-trivial-variable: log + use msg in DatabaseManager
  - cert-err33-c: NOLINT Boost.Asio error_code calls in Server WsListener
  - bugprone-throw-keyword-missing: NOLINTNEXTLINE base class ctor in hpp
T056-004: Document security_scan target in README.md under Security section

T057-001: Add docs/deployment.md — HTTP/WS TLS/port config, multi-tenant
  bootstrap (5-step guide), graceful shutdown (SIGTERM/SIGINT → Server::stop)
T057-002: Add embedded/Raspberry Pi deployment guidance in docs/deployment.md
  — min hardware table, cross-compilation profile, SQLite cache tuning,
  thread pool sizing for constrained environments

All 29 tests green: ctest --output-on-failure exits 0

+++ src/main/cpp/isched/backend/isched_CryptoUtils.cpp
@@ -143,7 +143,7 @@ std::string encrypt_secret(const std::string& plaintext,
         throw std::runtime_error("encrypt_secret: get GCM tag failed");
 
     // Wire format: nonce (12) | tag (16) | ciphertext
-    const std::size_t ct_len = static_cast<std::size_t>(enc_len + final_len);
+    const std::size_t ct_len = static_cast<std::size_t>(enc_len) + static_cast<std::size_t>(final_len);
     std::vector<unsigned char> blob;
     blob.reserve(12 + 16 + ct_len);
     blob.insert(blob.end(), nonce, nonce + 12);
